requirements.txt
seaborn==0.11.1
pycaret==2.3.3
pandas==1.2.5
matplotlib==3.4.2
gradio==2.2.8
Pillow==8.3.1
reportlab==3.6.1
